Make IPXE boot process more robust
It can happen, that ipmi commands get lost in the network. For power control we
already poll the ipmi device and set the desired state until it can be
confirmed.
However, the "force boot via PXE" did quite often not work as intended, and it
looks like a lost package or so. In order to avoid this in the future, we
should poll the ipmi device until the correct boot device is stored and set it
again until we can confirm success.

Signed-off-by: Michael Moese <mmoese@suse.de>