Change HA_CLUSTER_INIT value from 1/0 to yes/no
Originally HA_CLUSTER_INIT was a boolean setting and test modules
expected it to be set to something (usually 1, i.e., true), or not set
or set to 0 (false). This meant that it was not a setting that could be
used in YAML schedules' conditional scheduling, and a second setting
(HA_CLUSTER_SETUP set to init or join) had to be introduced. Since both
settings control the same, i.e., which job will initialize the HA
cluster and which node(s) will join to it, it was decided to consolidate
both settings into a single one. This commit updates the code so
available values for HA_CLUSTER_INIT are yes for the jobs where the HA
cluster is initialized, or different than yes for the rest of the jobs.