Add simple test for workload memory protection (WMP)
Workload memory protection prevents swapping for the cgroup configured
to contain the SAP processes.
This schedules a simple test, installing HANA, configuring WMP and also
running the simple WMP tests.
In order to test the required behaviour, we partition the system RAM as
follows: use a minimum of 32 GB for the cgroup, specify the amount in
the variable WMP_MEMORY_LOW and specify the remainder of the system
memory plus 1 GB in WMP_STRESS_MEM. There are no default values and no
automatic detection, but as this is a test to be executed on baremetal,
you should know the system specs of your target machine.

The tests take quite long to run, as we don't know when HANA is done
populating its memory or when stress-ng has built enough memory
pressure.

Signed-off-by: Michael Moese <mmoese@suse.de>